Include params are now optional.
The previous version of Manager getIncludeParams() did not return a ParamBag if no include parameters were set. This resulted in a ErrorException when a Transformers include function expected to be passed a ParamBag instance. This change makes getIncludeParams return an empty ParamBag so that parameters to a include function are optional rather than all or nothing.